A Note About Git Commit MessagesA Note About Git Commit MessagesA Note About Git Commit MessagesA Note About Git Commit MessagesA Note About Git Commit Messages
More detailed explanatory text, if necessary.  Wrap it to about 72 characters or so.  In some contexts, the first line is treated as the subject of an email and the rest of the text as the body.  The blank line separating the summary from the body is critical (unless you omit the body entirely); tools like rebase can get confused if you run the two together.